Gateway Service: Refactoring and quality of code
