+++ src/gateway/schemas/users.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    username: str
+    email: str
+
+
+class SensibleUser(User):
+    password: str
