+++ src/gateway/utils/logger.py
@@ -1,0 +1,5 @@
+import logging
+
+LOGGING = True
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
